feat(infra): add RPM package
While pg_idkit is not limited to linux distributions that support
RPMs, given that some users have requested an RPM package be produced
for pg_idkit, and there exists easy integration for publishing
RPMs (thakns to `cargo-generate-rpm`), it is easy to add some support
for RPMs

This commit adds an RPM package that is produced upon every major
release tag, utilizing cargo-generate-rpm.

While the pg_idkit does not currently go as far as to register the RPM
with any official entity, RPM releases will be made available vai GitHub releases.

Signed-off-by: vados <vados@vadosware.io>