refactor(docs): rework README
The documentation for pg_idkit isn't very tight/easy to use, despite
the ease of using `pgrx` and associated tooling/the extension itself.

This commit improves the README by making it terser, hiding
information that is unnecssary, and moving some contributor-specific
details and other niche information to a `docs` folder.

Signed-off-by: vados <vados@vadosware.io>