refactor(ops): dockerfile building
Until now, Postgres versions on which the Dockerfiles were built have
been locked to a given version, but that version *was not specified*
in the image name.

For example, the base-pkg-alpine3.18.Dockerfile was using Postgres
15.5 by default.

This commit refactors the Dockerfiles for both the base-pkg and
pg-idkit to take into account the Postgres base version that we build for.

Signed-off-by: vados <vados@vadosware.io>